Splitting EntityMapping into separate files
